Rewriting the site
I added the files from Bootstrap as I changed my html and css, using
Bootstrap. Also, I rewrote my code into a new file try.html so that I
could implement it in a better way, using columns and rows so that it
can be viewed on smaller devices and is responsive.